smart-contract-assistant/requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langserve>=0.0.30

# Free LLM (Ollama)
langchain-ollama>=0.1.0

# API Server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# UI
gradio>=4.0.0

# Document Processing
pymupdf>=1.23.0
python-docx>=1.1.0

# Vector Store
chromadb>=0.4.15

# Embeddings
sentence-transformers>=2.2.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
